anyMissing {Biobase}	R Documentation

Checks if there are any missing values in an object or not

Description

Checks if there are any missing values in an object or not.

Usage

anyMissing(x=NULL)

Arguments

x
A vector.

Details

The implementation of this method is optimized for both speed and memory.

Value

Returns TRUE if a missing value was detected, otherwise FALSE.

Author(s)

Henrik Bengtsson (http://www.braju.com/R/)

Examples


x <- rnorm(n=1000)
x[seq(300,length(x),by=100)] <- NA
stopifnot(anyMissing(x) == any(is.na(x)))

[Package Biobase version 2.42.0 Index]