channelNames {Biobase}	R Documentation

Retrieve and set channel names from object

Description

This generic function reports or updates the channels in an object.

Usage


channelNames(object, ...)
channelNames(object, ...) <- value

Arguments

object
An S4 object, typically derived from class eSet

value
Replacement value, either a character vector (to re-order existing channel names or a named character vector or list (to change channel names from the vector elements to the corresponding names).

...
Additional argument, not currently used.

Details

channelNames returns the names of the channels in a defined order. Change the order using the replacement method with a permuation of the channel names as value. Rename channels using the replacement method with a named list, where the vector elements are a permutation of the current channels, with corresponding names the new identifier for the channel.

Value

character.

Author(s)

Biocore

Examples


## channelNames default to alphabetical order of channels
obj <- NChannelSet(
           R=matrix(runif(100), 20, 5),
           G=matrix(-runif(100), 20, 5))
channelNames(obj)
channelNames(obj) <- c(Gn="G", Rd="R")   ## rename
channelNames(obj)
channelNames(obj) <- c("Rd", "Gn")       ## reorder
channelNames(obj)
all(assayData(obj)[["Gn"]] <= 0)
all(assayData(obj)[["Rd"]] >= 0)

[Package Biobase version 2.42.0 Index]