AssayData-class {Biobase}	R Documentation

Class "AssayData"

Description

Container class defined as a class union of list and environment. Designed to contain one or more matrices of the same dimension.

Methods

combine

signature(x = "AssayData", y = "AssayData"): This method uses cbind to create new AssayData elements that contain the samples of both arguments x and y.

Both AssayData arguments to combine must have the same collection of elements. The elements must have identical numbers of rows (features). The numerical contents of any columns (samples) present in the same element of different AssayData must be identical. The storageMode of the AssayData arguments must be identical, and the function returns an AssayData with storageMode matching the incoming mode. See also combine,eSet,eSet-method

featureNames

signature(object = "AssayData")

featureNames<-

signature(object = "AssayData", value = "ANY"): Return or set the feature names as a character vector. These are the row names of the AssayData elements. value can be a character or numeric vector; all entries must be unique.

sampleNames

signature(object = "AssayData")

sampleNames<-

signature(object = "AssayData", value="ANY"): Return or set the sample names. These are the column names of the the AssayData elements and the row names of phenoData. value can be a character or numeric vector.

storageMode

signature(object = "AssayData")

storageMode<-

signature(object = "AssayData", value="character"): Return or set the storage mode for the instance. value can be one of three choices: "lockedEnvironment", "environment", and "list". Environments offer a mechanism for storing data that avoids some of the copying that occurs when using lists. Locked environment help to ensure data integrity. Note that environments are one of the few R objects that are pass-by-reference. This means that if you modify a copy of an environment, you also modify the original. For this reason, we recommend using lockedEnvironment whenever possible.

Additional functions operating on AssayData include:

assayData[[name]]

Select element name from assayData.

assayDataNew(storage.mode = c("lockedEnvironment", "environment", "list"), ...)

Use storage.mode to create a new list or environment containing the named elements in ...

assayDataValidMembers(assayData, required)

Validate assayData, ensuring that the named elements required are present, matrices are of the same dimension, and featureNames (rownames) are consistent (identical or NULL) across entries.

assayDataElement(object, element)

See eSet-class

assayDataElementReplace(object, element, value, validate=TRUE)

See eSet-class

assayDataElementNames(object)

See eSet-class

Author(s)

Biocore

See Also

eSet-class ExpressionSet-class

[Package Biobase version 2.42.0 Index]