ExpressionSet {Biobase}	R Documentation

Class to Contain and Describe High-Throughput Expression Level Assays.

Description

Container for high-throughput assays and experimental metadata. ExpressionSet class is derived from eSet, and requires a matrix named exprs as assayData member.

Usage


## Instance creation
ExpressionSet(assayData,
    phenoData=annotatedDataFrameFrom(assayData, byrow=FALSE),
    featureData=annotatedDataFrameFrom(assayData, byrow=TRUE),
    experimentData=MIAME(), annotation=character(),
    protocolData=annotatedDataFrameFrom(assayData, byrow=FALSE),
    ...)

## Additional methods documented below

Arguments

assayData
A matrix of expression values, or an environment.

When assayData is a matrix, the rows represent probe sets (‘features’ in ExpressionSet parlance). Columns represent samples. When present, row names identify features and column names identify samples. Row and column names must be unique, and consistent with row names of featureData and phenoData, respectively. The assay data can be retrieved with exprs().

When assayData is an environment, it contains identically dimensioned matrices like that described in the previous paragraph. One of the elements of the environment must be named ‘exprs’; this element is returned with exprs().

phenoData
An optional AnnotatedDataFrame containing information about each sample. The number of rows in phenoData must match the number of columns in assayData. Row names of phenoData must match column names of the matrix / matricies in assayData.

featureData
An optional AnnotatedDataFrame containing information about each feature. The number of rows in featureData must match the number of rows in assayData. Row names of featureData must match row names of the matrix / matricies in assayData.

experimentData
An optional MIAME instance with meta-data (e.g., the lab and resulting publications from the analysis) about the experiment.

annotation
A character describing the platform on which the samples were assayed. This is often the name of a Bioconductor chip annotation package, which facilitated down-stream analysis.

protocolData
An optional AnnotatedDataFrame containing equipment-generated information about protocols. The number of rows and row names of protocolData must agree with the dimension and column names of assayData.

...
Additional arguments, passed to new("ExpressionSet", ...) and available for classes that extend ExpressionSet.

Extends

Directly extends class eSet.

Creating Objects

ExpressionSet instances are usually created through ExpressionSet().

Slots

Inherited from eSet:

assayData:

Contains matrices with equal dimensions, and with column number equal to nrow(phenoData). assayData must contain a matrix exprs with rows representing features (e.g., probe sets) and columns representing samples. Additional matrices of identical size (e.g., representing measurement errors) may also be included in assayData. Class:AssayData-class

phenoData:

See eSet

featureData:

See eSet

experimentData:

See eSet

annotation:

See eSet

protocolData:

See eSet

Methods

Class-specific methods.

as(exprSet,"ExpressionSet")

Coerce objects of exprSet-class to ExpressionSet

as(object,"data.frame")

Coerce objects of ExpressionSet-class to data.frame by transposing the expression matrix and concatenating phenoData

exprs(ExpressionSet), exprs(ExpressionSet,matrix)<-

Access and set elements named exprs in the AssayData-class slot.

esApply(ExpressionSet, MARGIN, FUN, ...)

'apply'-like function to conveniently operate on ExpressionSet objects. See esApply.

write.exprs(ExpressionSet)

Write expression values to a text file. It takes the same arguments as write.table

Derived from eSet:

updateObject(object, ..., verbose=FALSE)

Update instance to current version, if necessary. See updateObject and eSet

isCurrent(object)

Determine whether version of object is current. See isCurrent

isVersioned(object)

Determine whether object contains a 'version' string describing its structure . See isVersioned

assayData(ExpressionSet):

See eSet

sampleNames(ExpressionSet) and sampleNames(ExpressionSet)<-:

See eSet

featureNames(ExpressionSet), featureNames(ExpressionSet, value)<-:

See eSet

dims(ExpressionSet):

See eSet

phenoData(ExpressionSet), phenoData(ExpressionSet,value)<-:

See eSet

varLabels(ExpressionSet), varLabels(ExpressionSet, value)<-:

See eSet

varMetadata(ExpressionSet), varMetadata(ExpressionSet,value)<-:

See eSet

pData(ExpressionSet), pData(ExpressionSet,value)<-:

See eSet

varMetadata(ExpressionSet), varMetadata(ExpressionSet,value)

See eSet

experimentData(ExpressionSet),experimentData(ExpressionSet,value)<-:

See eSet

pubMedIds(ExpressionSet), pubMedIds(ExpressionSet,value)

See eSet

abstract(ExpressionSet):

See eSet

annotation(ExpressionSet), annotation(ExpressionSet,value)<-

See eSet

protocolData(ExpressionSet), protocolData(ExpressionSet,value)<-

See eSet

combine(ExpressionSet,ExpressionSet):

See eSet

storageMode(ExpressionSet), storageMode(ExpressionSet,character)<-:

See eSet

Standard generic methods:

initialize(ExpressionSet):

Object instantiation, used by new; not to be called directly by the user.

updateObject(ExpressionSet):

Update outdated versions of ExpressionSet to their current definition. See updateObject, Versions-class.

validObject(ExpressionSet):

Validity-checking method, ensuring that exprs is a member of assayData. checkValidity(ExpressionSet) imposes this validity check, and the validity checks of eSet.

makeDataPackage(object, author, email, packageName, packageVersion, license, biocViews, filePath, description=paste(abstract(object), collapse="\n\n"), ...)

Create a data package based on an ExpressionSet object. See makeDataPackage.

as(exprSet,ExpressionSet):

Coerce exprSet to ExpressionSet.

as(eSet,ExpressionSet):

Coerce the eSet portion of an object to ExpressionSet.

show(ExpressionSet)

See eSet

dim(ExpressionSet), ncol

See eSet

ExpressionSet[(index):

See eSet

ExpressionSet$, ExpressionSet$<-

See eSet

ExpressionSet[[i]], ExpressionSet[[i]]<-

See eSet

Author(s)

Biocore team

See Also

eSet-class, ExpressionSet-class.

Examples


# create an instance of ExpressionSet
ExpressionSet()

ExpressionSet(assayData=matrix(runif(1000), nrow=100, ncol=10))

# update an existing ExpressionSet
data(sample.ExpressionSet)
updateObject(sample.ExpressionSet)

# information about assay and sample data
featureNames(sample.ExpressionSet)[1:10]
sampleNames(sample.ExpressionSet)[1:5]
experimentData(sample.ExpressionSet)

# subset: first 10 genes, samples 2, 4, and 10
expressionSet <- sample.ExpressionSet[1:10,c(2,4,10)]

# named features and their expression levels
subset <- expressionSet[c("AFFX-BioC-3_at","AFFX-BioDn-5_at"),]
exprs(subset)

# samples with above-average 'score' in phenoData
highScores <- expressionSet$score > mean(expressionSet$score)
expressionSet[,highScores]

# (automatically) coerce to data.frame
lm(score~AFFX.BioDn.5_at + AFFX.BioC.3_at, data=subset)

[Package Biobase version 2.42.0 Index]