MIAME {Biobase}	R Documentation

Class for Storing Microarray Experiment Information

Description

Class MIAME covers MIAME entries that are not covered by other classes in Bioconductor. Namely, experimental design, samples, hybridizations, normalization controls, and pre-processing information. The MIAME class is derived from MIAxE.

Slots

name:

Object of class character containing the experimenter name

lab:

Object of class character containing the laboratory where the experiment was conducted

contact:

Object of class character containing contact information for lab and/or experimenter

title:

Object of class character containing a single-sentence experiment title

abstract:

Object of class character containing an abstract describing the experiment

url:

Object of class character containing a URL for the experiment

samples:

Object of class list containing information about the samples

hybridizations:

Object of class list containing information about the hybridizations

normControls:

Object of class list containing information about the controls such as house keeping genes

preprocessing:

Object of class list containing information about the pre-processing steps used on the raw data from this experiment

pubMedIds:

Object of class character listing strings of PubMed identifiers of papers relevant to the dataset

other:

Object of class list containing other information for which none of the above slots does not applies

Methods

Constructor methods:

MIAME():

MIAME(name = "", lab = "", contact = "", title = "", abstract = "", url = "", pubMedIds = "" , samples = "", hybridizations = list(), normControls = list(), preprocessing = list(), other = list()): Creates a new MIAME object with slots as defined above.

Class-specific methods:

abstract(MIAME):

An accessor function for abstract.

combine(MIAME,MIAME):

Combine two objects of MIAME-class, issuing warnings when ambiguities encountered.

expinfo(MIAME):

An accessor function for name, lab, contact, title, and url.

hybridizations(MIAME):

An accessor function for hybridizations.

normControls(MIAME):

An accessor function for normControls.

notes(MIAME), notes(MIAME) <- value:

Accessor functions for other. notes(MIAME) <- character appends character to notes; use notes(MIAME) <- list to replace the notes entirely.

otherInfo(MIAME):

An accessor function for other.

preproc(MIAME):

An accessor function for preprocessing.

pubMedIds(MIAME), pubMedIds(MIAME) <- value:

Accessor function for pubMedIds.

samples(MIAME):

An accessor function for samples.

Standard generic methods:

updateObject(object, ..., verbose=FALSE)

Update instance to current version, if necessary. See updateObject

isCurrent(object)

Determine whether version of object is current. See isCurrent

isVersioned(object)

Determine whether object contains a 'version' string describing its structure . See isVersioned

show(MIAME):

Renders information about the MIAME information

Author(s)

Rafael A. Irizarry

References

http://www.mged.org/Workgroups/MIAME/miame_1.1.html

See Also

class:characterORMIAME, read.MIAME

[Package Biobase version 2.42.0 Index]