MultiSet {Biobase}	R Documentation

Class to Contain and Describe High-Throughput Expression Level Assays.

Description

Container for high-throughput assays and experimental metadata. MutliSet is derived from eSet-class. MultiSet differs from ExpressionSet-class because MultiSet can contain any element(s) in assayData (ExpressionSet must have an element named exprs).

Extends

Directly extends class eSet.

Creating Objects

new('MultiSet', phenoData = [AnnotatedDataFrame], experimentData = [MIAME], annotation = [character], protocolData = [AnnotatedDataFrame], ...)

updateOldESet(oldESet,"MultiSet")

MultiSet instances are usually created through new("MultiSet", ...). The ... arguments to new are matrices of expression data (with features corresponding to rows and samples to columns), phenoData, experimentData, annotation, and protocolData. phenoData, experimentData, annotation, and protocolData can be missing, in which case they are assigned default values.

updateOldESet will take a serialized instance (e.g., saved to a disk file with save object created with earlier definitions of the eSet-class, and update the object to MultiSet. Warnings are issued when direct translation is not possible; incorrectly created oldESet instances may not be updated.

Slots

Inherited from eSet:

assayData:

Contains zero or more matrices with equal dimensions, and with column number equal to nrow(phenoData). Each matrix in assayData has rows representing features (e.g., reporters) and columns representing samples. Class:AssayData-class

phenoData:

See eSet-class

experimentData:

See eSet-class

annotation:

See eSet-class

protocolData:

See eSet-class

Methods

Class-specific methods: none

Derived from eSet-class:

updateObject(object, ..., verbose=FALSE)

Update instance to current version, if necessary. See updateObject and eSet

isCurrent(object)

Determine whether version of object is current. See isCurrent

isVersioned(object)

Determine whether object contains a 'version' string describing its structure . See isVersioned

sampleNames(MultiSet) and sampleNames(MultiSet)<-:

See eSet-class

featureNames(MultiSet), featureNames(MultiSet, value)<-:

See eSet-class

dims(MultiSet):

See eSet-class

phenoData(MultiSet), phenoData(MultiSet,value)<-:

See eSet-class

varLabels(MultiSet), varLabels(MultiSet, value)<-:

See eSet-class

varMetadata(MultiSet), varMetadata(MultiSet,value)<-:

See eSet-class

pData(MultiSet), pData(MultiSet,value)<-:

See eSet-class

varMetadata(MultiSet), varMetadata(MultiSet,value)

See eSet-class

experimentData(MultiSet),experimentData(MultiSet,value)<-:

See eSet-class

pubMedIds(MultiSet), pubMedIds(MultiSet,value)

See eSet-class

abstract(MultiSet):

See eSet-class

annotation(MultiSet), annotation(MultiSet,value)<-

See eSet-class

protocolData(MultiSet), protocolData(MultiSet,value)<-

See eSet-class

combine(MultiSet,MultiSet):

See eSet-class

storageMode(eSet), storageMode(eSet,character)<-:

See eSet-class

Standard generic methods:

initialize(MultiSet):

Object instantiation, used by new; not to be called directly by the user.

validObject(MultiSet):

Validity-checking method, ensuring that all elements of assayData are matricies with equal dimensions.

as(eSet,MultiSet):

Coerce the eSet portion of an object to MultiSet.

show(MultiSet)

See eSet-class

dim(MultiSet), ncol

See eSet-class

MultiSet[(index):

See eSet-class

MultiSet$, MultiSet$<-

See eSet-class

Author(s)

Biocore team

See Also

eSet-class, ExpressionSet-class

Examples


# create an instance of ExpressionSet
new("MultiSet")

[Package Biobase version 2.42.0 Index]