SnpSet {Biobase}	R Documentation

Class to Contain Objects Describing High-Throughput SNP Assays.

Description

Container for high-throughput assays and experimental metadata. SnpSet class is derived from eSet, and requires matrices call, callProbability as assay data members.

Extends

Directly extends class eSet.

Creating Objects

new('SnpSet', phenoData = [AnnotatedDataFrame], experimentData = [MIAME], annotation = [character], protocolData = [AnnotatedDataFrame], call = [matrix], callProbability = [matrix], ...)

SnpSet instances are usually created through new("SnpSet", ...). Usually the arguments to new include call (a matrix of genotypic calls, with features (SNPs) corresponding to rows and samples to columns), phenoData, experimentData, annotation, and protocolData. phenoData, experimentData, annotation and protocolData can be missing, in which case they are assigned default values.

Slots

Inherited from eSet:

assayData:

Contains matrices with equal dimensions, and with column number equal to nrow(phenoData). assayData must contain a matrix call with rows representing features (e.g., SNPs) and columns representing samples, and a matrix callProbability describing the certainty of the call. The content of call and callProbability are not enforced by the class. Additional matrices of identical size may also be included in assayData. Class:AssayData-class

phenoData:

See eSet

experimentData:

See eSet

annotation:

See eSet

protocolData:

See eSet

Methods

Class-specific methods:

snpCall(SnpSet), snpCall(SnpSet,matrix)<-

Access and set elements named call in the AssayData slot.

exprs(SnpSet), exprs(SnpSet,matrix)<-

Synonym for snpCall.

snpCallProbability(SnpSet), snpCallProbability<-(SnpSet,matrix)<-

Access and set elements named callProbability in the AssayData slot.

Derived from eSet:

updateObject(object, ..., verbose=FALSE)

Update instance to current version, if necessary. See updateObject and eSet

isCurrent(object)

Determine whether version of object is current. See isCurrent

isVersioned(object)

Determine whether object contains a 'version' string describing its structure . See isVersioned

sampleNames(SnpSet) and sampleNames(SnpSet)<-:

See eSet

featureNames(SnpSet), featureNames(SnpSet, value)<-:

See eSet

dims(SnpSet):

See eSet

phenoData(SnpSet), phenoData(SnpSet,value)<-:

See eSet

varLabels(SnpSet), varLabels(SnpSet, value)<-:

See eSet

varMetadata(SnpSet), varMetadata(SnpSet,value)<-:

See eSet

pData(SnpSet), pData(SnpSet,value)<-:

See eSet

varMetadata(SnpSet), varMetadata(SnpSet,value)

See eSet

experimentData(SnpSet),experimentData(SnpSet,value)<-:

See eSet

pubMedIds(SnpSet), pubMedIds(SnpSet,value)

See eSet

abstract(SnpSet):

See eSet

annotation(SnpSet), annotation(SnpSet,value)<-

See eSet

protocolData(SnpSet), protocolData(SnpSet,value)<-

See eSet

combine(SnpSet,SnpSet):

See eSet

storageMode(eSet), storageMode(eSet,character)<-:

See eSet

Standard generic methods:

initialize(SnpSet):

Object instantiation, used by new; not to be called directly by the user.

validObject(SnpSet):

Validity-checking method, ensuring that call and callProbability is a member of assayData. checkValidity(SnpSet) imposes this validity check, and the validity checks of eSet.

show(SnpSet)

See eSet

dim(SnpSet), ncol

See eSet

SnpSet[(index):

See eSet

SnpSet$, SnpSet$<-

See eSet

Author(s)

Martin Morgan, V.J. Carey, after initial design by R. Gentleman

See Also

eSet, ExpressionSet

[Package Biobase version 2.42.0 Index]