Versions {Biobase}	R Documentation

Class "Versions"

Description

A class to record version number information. This class is used to report versions; to add version information to your own class, use Versioned-class.

Methods

The following are defined; package developers may write additional methods.

new("Versions", ...)

Create a new Versions-class instance, perhaps with named version elements (the contents of ...) added. Named elements of versions are character strings that can be coerced using package_version, or package_version instances, Versions-class objects.

object["id"]

Obtain version information "id" from object.

object["id"] <- value

Create or update version information "id" on instance object.

object[["id"]]

Obtain version information "id" from object. The result is a list of integers, corresponding to entries in the version string.

object[["id"]] <- value

Create or update version information "id" on instance object.

object$id

Obtain version information "id" from object.The result is a list of integers, corresponding to entries in the version string.

object$id <- value

Create or update version information "id" on instance object.

show(object)

Display version information.

updateObject(object)

Update object to the current Versions-class representation. Note that this does not update another class that uses Versions-class to track the class version.

as(object, "character")

Convert object to character representation, e.g., 1.0.0

object1 < object2

Compare object1 and object2 using version class information. Symbols in addition to < are admissible; see ?Ops

Author(s)

Biocore

See Also

classVersion isCurrent isVersioned

Examples



obj <- new("Versions", A="1.0.0")
obj

obj["A"] <- "1.0.1"
obj
obj["B"] <- "2.0"
obj

obj1 <- obj
obj1["B"] <- "2.0.1"

obj1 == obj
obj1["B"] > "2.0.0"
obj["B"] == "2.0" # TRUE!


[Package Biobase version 2.42.0 Index]