dims {Biobase}	R Documentation

Retrieve dimensions of all elements in a list or environment

Description

This function returns the dimensions of element members in lists or environments such as AssayData-class.

Usage

dims(object)

Arguments

object
List or environment object containing one or several matrices

Value

matrix of row and column dimensions, (in rows) for each element in object (columns).

Author(s)

Biocore

See Also

eSet-class

[Package Biobase version 2.24.0 Index]