exprs {Biobase}	R Documentation

Retrieve expression data from eSets.

Description

These generic functions access the expression and error measurements of assay data stored in an object derived from the eSet-class.

Usage


exprs(object)
exprs(object) <- value
se.exprs(object)
se.exprs(object) <- value

Arguments

object
Object derived from class eSet.

value
Matrix with rows representing features and columns samples.

Value

exprs returns a (usually large!) matrix of expression values; se.exprs returns the corresponding matrix of standard errors, when available.

Author(s)

Biocore

See Also

eSet-class, ExpressionSet-class, SnpSet-class

[Package Biobase version 2.42.0 Index]