getPkgVigs {Biobase}	R Documentation

List Vignette Files for a Package

Description

This function will return a listing of all vignettes stored in a package's doc directory.

Usage


getPkgVigs(package = NULL)

Arguments

package
A character vector of packages to search or NULL. The latter is for all attached packages (in search()).

Value

A data.frame with columns package, filename, title.

Author(s)

Jeff Gentry, modifications by Wolfgang Huber.

See Also

openVignette

Examples


 z <- getPkgVigs()
 z # and look at them

[Package Biobase version 2.42.0 Index]