isUnique {Biobase}	R Documentation

Determine Unique Elements

Description

Determines which elements of a vector occur exactly once.

Usage

isUnique(x)

Arguments

x
a vector

Value

A logical vector of the same length as x, in which TRUE indicates uniqueness.

Author(s)

Wolfgang Huber

See Also

unique,duplicated.

Examples

x <- c(9:20, 1:5, 3:7, 0:8)
isUnique(x)

[Package Biobase version 2.70.0 Index]