isVersioned {Biobase}	R Documentation

Determine whether object or class contains versioning information

Description

This generic function checks to see whether Versioned-class information is present. When the argument to isVersioned is a character string, the prototype of the class corresponding to the string is consulted.

By default, isVersioned has the following behaviors:

isVersioned(Versioned-instance)

Returns TRUE when the instance have version information.

isCurrent("class-name")

Returns TRUE when the named class extends Versioned-class.

isVersioned(ANY)

Returns FALSE

Usage


isVersioned(object)

Arguments

object
Object or class name to check for version information, as described above.

Value

isVersioned returns a logical indicating whether version information is present.

Author(s)

Biocore team

See Also

Versions-class

Examples



obj <- new("VersionedBiobase")
isVersioned(obj)

isVersioned(1:10) # FALSE

A <- setClass("A", contains="VersionedBiobase",
         prototype=prototype(new("VersionedBiobase", versions=c(A="1.0.0"))))
a <- A()
isVersioned(a)

removeClass("A")

[Package Biobase version 2.42.0 Index]