makeDataPackage {Biobase}	R Documentation

Make an R package from a data object

Description

This generic creates a valid R package from an R data object.

Usage


makeDataPackage(object, author, email,
                packageName=deparse(substitute(object)),
                packageVersion=package_version("1.0.0"),
                license="Artistic-2.0",
                biocViews="ExperimentData",
                filePath=tempdir(),
                ...)

Arguments

object
An instance of an R data object.

author
The author, as a character string.

email
A valid email address for the maintainer, as a character string.

packageName
The name of the package, defaults to the name of the object instance.

packageVersion
The version number, as a character string.

license
The license, as a character string.

biocViews
A character vector of valid biocViews views.

filePath
The location to create the package.

...
Additional arguments to specific methods.

Details

The function makes use of various tools in R and Bioconductor to automatically generate the source files for a valid R package.

Value

The return value is that from a call to link{createPackage} which is invoked once the default arguments are set up. The data instance is stored in the data directory with a name the same as that of the resulting package.

Note

Developers implementing derived methods might force correct package name evaluation by including 'packageName' in any callNextMethod().

Author(s)

R. Gentleman

See Also

createPackage

Examples


 data(sample.ExpressionSet)
 ## package created in tempdir()
 s1 <- makeDataPackage(sample.ExpressionSet,
                       author = "Foo Author",
                       email = "foo@bar",
                       packageName = "FooBarPkg",
                       packageVersion = "1.0.0")

[Package Biobase version 2.42.0 Index]