note {Biobase}	R Documentation

Informational Messages

Description

Generates an informational message that corresponds to its argument(s). Similar to warning() except prefaced by "Note:" instead of "Warning message:".

Usage


note(...)

Arguments

...
character vectors (which are pasted together) or NULL

Details

This function essentially cat()'s the created string to the screen. It is intended for messages to the user that are deemed to be 'informational', as opposed to warnings, etc.

Author(s)

Jeff Gentry

See Also

warning,stop

Examples


   note("This is an example of a note")

[Package Biobase version 2.42.0 Index]