openVignette {Biobase}	R Documentation

Open a Vignette or Show Vignette Selection Menu

Description

Using the data returned by vignette this function provides a simple easy to use interface for opening vignettes.

Usage


openVignette(package=NULL)

Arguments

package
character string indicating the package to be used.

Details

If package is NULL then all packages are scanned for vignettes. The list of vignettes is presented to the user via the menu command. The user may select one of the vignettes to be opened in a PDF viewer.

Value

No value is returned; this function is run entirely for the side effect of opening the pdf document in the PDF viewer.

Author(s)

R. Gentleman

See Also

vignette, openPDF, menu, getPkgVigs

Examples


  if( interactive() )
    openVignette("Biobase")

[Package Biobase version 2.42.0 Index]