read.MIAME {Biobase}	R Documentation

Read MIAME Information into an Instance of Class 'MIAME'

Description

Reads MIAME information from a file or using a widget.

Usage


read.MIAME(filename = NULL, widget = getOption("BioC")$Base$use.widgets, ...)

Arguments

filename
Filename from which to read MIAME information.

widget
Logical. If TRUE and a filename is not given, a widget is used to enter information.

...
Further arguments to scan.

Details

Notice that the MIAME class tries to cover the MIAME entries that are not covered by other classes in Bioconductor. Namely, experimental design, samples, hybridizations, normalization controls, and pre-processing information.

The function scan is used to read. The file must be a flat file with the different entries for the instance of MIAME class separated by carriage returns. The order should be: name, lab, contact, title, abstract, and url.

Alternatively a widget can be used.

Value

An object of class MIAME.

Author(s)

Rafael Irizarry <rafa@jhu.edu>

See Also

MIAME,tkMIAME

Examples


miame <- read.MIAME(widget=FALSE) ##creates an empty instance
show(miame)

[Package Biobase version 2.42.0 Index]