readExpressionSet {Biobase}	R Documentation

Read 'ExpressionSet'

Description

Create an instance of class ExpressionSet by reading data from files. ‘widget’ functionality is not implemented for readExpressionSet.

Usage


readExpressionSet(exprsFile,
      phenoDataFile,
      experimentDataFile,
      notesFile,
      path,
      annotation,
      ## arguments to read.* methods 
      exprsArgs=list(sep=sep, header=header, row.names=row.names,
        quote=quote, ...),
      phenoDataArgs=list(sep=sep, header=header, row.names=row.names,
        quote=quote, stringsAsFactors=stringsAsFactors, ...),
      experimentDataArgs=list(sep=sep, header=header,
        row.names=row.names, quote=quote,
        stringsAsFactors=stringsAsFactors, ...),
      sep = "\t", header = TRUE, quote = "", stringsAsFactors = FALSE,
      row.names = 1L,
      ## widget
      widget = getOption("BioC")$Base$use.widgets,
      ...)

Arguments

exprsFile
(character) File or connection from which to read expression values. The file should contain a matrix with rows as features and columns as samples. read.table is called with this as its file argument and further arguments given by exprsArgs.

phenoDataFile
(character) File or connection from which to read phenotypic data. read.AnnotatedDataFrame is called with this as its file argument and further arguments given by phenoDataArgs.

experimentDataFile
(character) File or connection from which to read experiment data. read.MIAME is called with this as its file argument and further arguments given by experimentDataArgs.

notesFile
(character) File or connection from which to read notes; readLines is used to input the file.

path
(optional) directory in which to find all the above files.

annotation
(character) A single character string indicating the annotation associated with this ExpressionSet.

exprsArgs
A list of arguments to be used with read.table when reading in the expression matrix.

phenoDataArgs
A list of arguments to be used (with read.AnnotatedDataFrame) when reading the phenotypic data.

experimentDataArgs
A list of arguments to be used (with read.MIAME) when reading the experiment data.

sep, header, quote, stringsAsFactors, row.names
arguments used by the read.table-like functions.

widget
A boolean value indicating whether widgets can be used. Widgets are NOT yet implemented for read.AnnotatedDataFrame.

...
Further arguments that can be passed on to the read.table-like functions.

Details

Expression values are read using the read.table function. Phenotypic data are read using the read.AnnotatedDataFrame function. Experiment data are read using the read.MIAME function. Notes are read using the readLines function. The return value must be a valid ExpressionSet. Only the exprsFile argument is required.

Value

An instance of the ExpressionSet class.

Author(s)

Martin Morgan <mtmorgan@fhcrc.org>

See Also

ExpressionSet for additional methods.

Examples



exprsFile = system.file("extdata", "exprsData.txt", package="Biobase")
phenoFile = system.file("extdata", "pData.txt", package="Biobase")

## Read ExpressionSet with appropriate parameters
obj = readExpressionSet(exprsFile, phenoFile, sep = "\t", header=TRUE)
obj


[Package Biobase version 2.42.0 Index]