rowQ {Biobase}	R Documentation

A function to compute empirical row quantiles.

Description

This function computes the requested quantile for each row of a matrix, or of an ExpressionSet.

Usage


rowQ(imat, which)
rowMax(imat)
rowMin(imat)

Arguments

imat
Either a matrix or an ExpressionSet.

which
An integer indicating which order statistic should be returned.

Details

rowMax and rowMin simply call rowQ with the appropriate argument set.

The argument which takes values between 1, for the minimum per row, and ncol(imat), for the maximum per row.

Value

A vector of length equal to the number of rows of the input matrix containing the requested quantiles.

Author(s)

R. Gentleman

See Also

rowMedians. rowMeans() in colSums().

Examples


 data(sample.ExpressionSet)
 rowMin(sample.ExpressionSet)
 rowQ(sample.ExpressionSet, 4)

[Package Biobase version 2.42.0 Index]