snpCall {Biobase}	R Documentation

Get and retrieve SNP call and call probability data.

Description

These generic functions access the calls and call probabilities stored in objects.

Usage


snpCall(object, ...)
snpCall(object, ...) <- value
snpCallProbability(object, ...)
snpCallProbability(object, ...) <- value

Arguments

object
Object, possibly derived from class SnpSet.

value
Matrix with rows representing SNP calls or call probabilities and columns samples.

...
Additional arguments available to methods.

Value

snpCall returns a matrix of SNP calls; snpCallProbability returns the corresponding matrix of standard errors, when available.

Author(s)

Biocore

See Also

SnpSet-class

[Package Biobase version 2.42.0 Index]