strbreak {Biobase}	R Documentation

Break Character Strings to Fit Width

Description

Inserts line breaks (collapse) into input character strings. The main intention of this function is to prepare long strings for printing, so the output is not wider than width.

Usage


  strbreak(x, width=getOption("width"), exdent=2, collapse="\n")

Arguments

x
a character vector

width
a positive integer giving the width of the output.

exdent
a positive integer specifying the indentation of subsequent lines after the first line.

collapse
a character. This is inserted to break lines.

Author(s)

Wolfgang Huber http://www.ebi.ac.uk/huber

See Also

strwrap, substring

Examples


longString = paste(rep(LETTERS, 10), collapse="", sep="")
cat(strbreak(longString))

[Package Biobase version 2.42.0 Index]