updateOldESet {Biobase}	R Documentation

Update previously created eSet object to current eSet structure

Description

This function updates eSet objects created in previous versions of Biobase to the current class structure. Warnings indicate when coercions change how data in the from object are altered. If the from object was not a valid object of the original eSet class, then updateOldESet may fail.

Usage


updateOldESet(from, toClass, ...)

Arguments

from
Object created using a previous version of the eSet class.

toClass
Character string identifying new class, e.g., "ExpressionSet"

...
Additional arguments passed to the initialization method for class toClass

Value

Valid object of class toClass.

Author(s)

Biocore

See Also

eSet-class, ExpressionSet-class, SnpSet-class

Examples


## Not run: 
updateOldESet(oldESet, "ExpressionSet")

## End(Not run)

[Package Biobase version 2.42.0 Index]