userQuery {Biobase}	R Documentation

A function to query the user for input

Description

This function will output a given message and seek a response from the user, repeating the message until the input is from a valid set provided by the code.

Usage


userQuery(msg, allowed = c("y", "n"), default = "n", case.sensitive = FALSE)

Arguments

msg
The output message

allowed
Allowed input from the user

default
Default response if called in batch mode

case.sensitive
Is the response case sensitive? Defaults to FALSE

Value

The input from the user

Author(s)

Jeff Gentry

[Package Biobase version 2.42.0 Index]