Biobase-package {Biobase}	R Documentation

Biobase Package Overview

Description

Biobase Package Overview

Details

Important data classes: ExpressionSet, AnnotatedDataFrame MIAME. Full help on methods and associated functions is available from within class help pages.

Additional data classes: eSet, MIAxE, MultiSet. Additional manipulation and data structuring classes: Versioned, VersionedBiobase, aggregator, container.

Vignette routines: openVignette, getPkgVigs, openPDF.

Package manipulation functions: createPackage and package.version

Data sets: aaMap, sample.ExpressionSet, geneData.

Introductory information is available from vignettes, type openVignette().

Full listing of documented articles is available in HTML view by typing help.start() and selecting Biobase package from the Packages menu or via library(help="Biobase").

Author(s)

O. Sklyar

[Package Biobase version 2.54.0 Index]