abstract {Biobase}	R Documentation

Retrieve Meta-data from eSets and ExpressionSets.

Description

These generic functions access generic data, abstracts, PubMed IDs and experiment data from instances of the eSet-class or ExpressionSet-class.

Usage

abstract(object)
pubMedIds(object)
pubMedIds(object) <- value
experimentData(object)
experimentData(object) <- value

Arguments

object
Object, possibly derived from eSet-class or MIAME-class

value
Value to be assigned; see class of object (e.g., eSet-class) for specifics.

Value

abstract returns a character vector containing the abstract (as in a published paper) associated with object.

pubMedIds returns a character vector of PUBMED IDs associated with the experiment.

experimentData returns an object representing the description of an experiment, e.g., an object of MIAME-class

Author(s)

Biocore

See Also

ExpressionSet-class, eSet-class, MIAME-class

[Package Biobase version 2.54.0 Index]