addVigs2WinMenu {Biobase}	R Documentation

Add Menu Items to an Existing/New Menu of Window

Description

This function adds a menu item for a package's vignettes.

Usage

addVigs2WinMenu(pkgName)

Arguments

pkgName
pkgName - a character string for the name of an R package

Details

The original functions addVig2Menu, addVig4Win, addVig4Unix, addNonExisting, addPDF2Vig have been replaced by addVigs2WinMenu, please use those instead.

Value

The functions do not return any value.

Author(s)

Jianhua Zhang and Jeff Gentry

Examples

    # Only works for windows now
    if(interactive() && .Platform$OS.type == "windows" &&
             .Platform$GUI == "Rgui"){
        addVigs2WinMenu("Biobase")
    }

[Package Biobase version 2.70.0 Index]