assayData {Biobase}	R Documentation

Retrieve assay data from eSets and ExpressionSets.

Description

This generic function accesses assay data stored in an object derived from the eSet or ExpressionSet class.

Usage

assayData(object)
assayData(object) <- value

Arguments

object
Object derived from class eSet

value
Named list or environment containing one or more matrices with identical dimensions

Value

assayData applied to eSet-derived classes returns a list or environment; applied to ExpressionSet, the method returns an environment. See the class documentation for specific details.

Author(s)

Biocore

See Also

eSet-class, ExpressionSet-class, SnpSet-class

[Package Biobase version 2.54.0 Index]