MIAxE {Biobase}	R Documentation

MIAxE objects

Description

The MIAxE virtual class is a general container for storing experiment metadata. Information such as experimental design, samples, normalization methods and pre-processing information can be stored in these objets.

The MIAxE class is virtual and MIAxE objects cannot be instantiated directly. The following classes derive directly from the MIAxE class: MIAME.

Slots

Introduced in MIAxE:

.__classVersion__:

A Versions object describing the MIAxE version number. Intended for developer use.

Methods

Standard generic methods:

show(object)

Informatively display object contents.

Author(s)

Biocore team

See Also

Related classes MIAME-class, ExpressionSet-class. Derived classes MIAME-class.

Examples

# Create a new class
MyData <- setClass("MyData", contains="MIAxE")
MyData()

# MIAxE objects cannot be instantiated directly
try(new("MIAxE"))

[Package Biobase version 2.70.0 Index]