classVersion {Biobase}	R Documentation

Retrieve information about versioned classes

Description

These generic functions return version information for classes derived from Versioned-class, or VersionsNull-class for unversioned objects. The version information is an object of Versions-class.

By default, classVersion has the following behaviors:

classVersion(Versioned-instance)

Returns a Versions-class object obtained from the object.

classVersion{"class"}

Consults the definition of class and return the current version information, if available.

classVersion(ANY)

Return a VersionsNull-class object to indicate no version information available.

By default, the classVersion<- method has the following behavior:

classVersion(Versioned-instance)["id"] <- value

Assign (update or add) value to Versions-instance. value is coerced to a valid version description. see Versions-class for additional access methods.

Usage

classVersion(object)
classVersion(object) <- value

Arguments

object
Object whose version is to be determined, as described above.

value
Version-class object to assign to object of Versioned-class object.

Value

classVersion returns an instance of Versions-class

Author(s)

Biocore team

See Also

Versions-class

Examples

obj <- new("VersionedBiobase")

classVersion(obj)
classVersion(obj)["Biobase"]
classVersion(1:10) # no version
classVersion("ExpressionSet") # consult ExpressionSet prototype

classVersion(obj)["MyVersion"] <- "1.0.0"
classVersion(obj)

[Package Biobase version 2.70.0 Index]