contents {Biobase}	R Documentation

Function to retrieve contents of environments

Description

The contents method is used to retrieve the values stored in an environment.

Usage

contents(object, all.names)

Arguments

object
The environment (data table) that you want to get all contents from

all.names
a logical indicating whether to copy all values in as.list.environment

Value

A named list is returned, where the elements are the objects stored in the environment. The names of the elements are the names of the objects.

The all.names argument is identical to the one used in as.list.environment.

Author(s)

R. Gentleman

See Also

as.list.environment

Examples

  z <- new.env()
  multiassign(letters, 1:26, envir=z)
  contents(z)

[Package Biobase version 2.70.0 Index]