copyEnv {Biobase}	R Documentation

List-Environment interactions

Description

These functions can be used to make copies of environments, or to get/assign all of the objects inside of an environment.

Usage

copyEnv(oldEnv, newEnv, all.names=FALSE)

Arguments

oldEnv
An environment to copy from

newEnv
An environment to copy to. If missing, a new environment with the same parent environment as oldEnv.

all.names
Whether to retrieve objects with names that start with a dot.

Details

copyEnv: This function will make a copy of the contents from oldEnv and place them into newEnv.

Author(s)

Jeff Gentry and R. Gentleman

See Also

environment, as.list

Examples

   z <- new.env(hash=TRUE, parent=emptyenv(), size=29L)
   multiassign(c("a","b","c"), c(1,2,3), z)

   a <- copyEnv(z)
   ls(a)

[Package Biobase version 2.70.0 Index]