data:sample.MultiSet {Biobase}	R Documentation

Data set of class 'MultiSet'

Description

The expression data are real but anonymized. The data are from an experiment that used Affymetrix U95v2 chips. The data were processed by dChip and then exported to R for analysis.

The phenoData, standard error estimates, and description data are fake.

Usage

data(sample.MultiSet)

Format

The data for 4 cases, labeled a to d and 500 genes. Each case has five covariates: SlideNumber: number; FileName: name; Cy3: genotype labeled Cy3; Cy5: genotype labeled Cy5; Date: date.

Examples

data(sample.MultiSet)

[Package Biobase version 2.70.0 Index]