featureData {Biobase}	R Documentation

Retrieve information on features recorded in eSet-derived classes.

Description

These generic functions access feature data (experiment specific information about features) and feature meta-data (e.g., descriptions of feature covariates).

Usage

featureData(object)
featureData(object) <- value
fData(object)
fData(object) <- value
fvarLabels(object)
fvarLabels(object) <- value
fvarMetadata(object)
fvarMetadata(object) <- value

Arguments

object
Object, possibly derived from eSet-class or AnnotatedDataFrame-class.

value
Value to be assigned to corresponding object.

Value

featureData returns an object containing information on both variable values and variable meta-data. fvarLabels returns a character vector of measured variable names. fData returns a data frame with features as rows, variables as columns. fvarMetadata returns a data frame with variable names as rows, description tags (e.g., unit of measurement) as columns.

Author(s)

Biocore

See Also

eSet, ExpressionSet

[Package Biobase version 2.70.0 Index]