openPDF {Biobase}	R Documentation

Open PDF Files in a Standard Viewer

Description

Displays the specified PDF file.

Usage

openPDF(file, bg=TRUE)

Arguments

file
A character string, indicating the file to view

bg
Should the pdf viewer be opened in the background.

Details

Currently this function works on Windows and Unix platforms. Under Windows, whatever program is associated with the file extension will be used. Under Unix, the function will use the program named in the option "pdfviewer" (see help(options) for information on how this is set.)

The bg argument is only interpreted on Unix.

Value

This function is executed for its side effects. The specified PDF file is opened in the PDF viewer and TRUE is returned.

Author(s)

Jeff Gentry

Examples

## Not run: openPDF("annotate.pdf")

[Package Biobase version 2.70.0 Index]