package.version {Biobase}	R Documentation

Report Version of a Package

Description

Will report the version number of a requested installed package

Usage

package.version(pkg, lib.loc = NULL)

Arguments

pkg
The name of the package

lib.loc
a character vector describing the location of R library trees to search through, or ‘NULL’. The default value of ‘NULL’ corresponds to all libraries currently known.

Details

This function is a convenience wrapper around package.description, and will report simply the version number of the requested package. If the package does not exist or if the DESCRIPTION file can not be read, then an error will be thrown.

Value

A character string reporting the version number.

Author(s)

Jeff Gentry

See Also

package.description

Examples

  package.version("Biobase")

[Package Biobase version 2.70.0 Index]