phenoData {Biobase}	R Documentation

Retrieve information on experimental phenotypes recorded in eSet and ExpressionSet-derived classes.

Description

These generic functions access the phenotypic data (e.g., covariates) and meta-data (e.g., descriptions of covariates) associated with an experiment.

Usage

phenoData(object)
phenoData(object) <- value
varLabels(object)
varLabels(object) <- value
varMetadata(object)
varMetadata(object) <- value
pData(object)
pData(object) <- value

Arguments

object
Object, possibly derived from eSet-class or AnnotatedDataFrame.

value
Value to be assigned to corresponding object.

Value

phenoData returns an object containing information on both variable values and variable meta-data. varLabels returns a character vector of measured variables. pData returns a data frame with samples as rows, variables as columns. varMetadata returns a data frame with variable names as rows, description tags (e.g., unit of measurement) as columns.

Author(s)

Biocore

See Also

eSet-class, ExpressionSet-class, SnpSet-class

[Package Biobase version 2.54.0 Index]