testBioCConnection {Biobase}	R Documentation

A function to check internet connectivity to Bioconductor

Description

This function will attempt to determine if the user has internet connectivity to the Bioconductor website. This is useful in many situations dealing with code that uses automated downloads and other such things.

Usage

testBioCConnection()

Value

TRUE if a connection is possible, FALSE if not.

Author(s)

Jeff Gentry

Examples

 z <- testBioCConnection()

[Package Biobase version 2.70.0 Index]