updateObjectTo {Biobase}	R Documentation

Update an object to the class definition of a template

Description

The updateObjectTo generic function returns an instance of object updated to the class definition of template.

It requires that the class of the returned object be the same as the class of the template argument, and that the object is valid. Usually, updating proceeds by modifying slots in template with information from object, and returning template. Use as to coerce an object from one type to another; updateObjectTo might be useful to update a virtual superclass. By default, updateObjectTo has the following behavior:

updateObjectTo(ANY-object,ANY-template)

Attempt as(ANY-object,class(ANY-template)).

Usage

updateObjectTo(object, template, ..., verbose=FALSE)

Arguments

object
Object to be updated.

template
Instance representing a template for updating object.

...
Additional arguments, for use in specific update methods.

verbose
A logical, indicating whether information about the update should be reported. Use message to report this.

Value

updateObjectTo returns a valid instance of template.

Author(s)

Biocore team

See Also

updateObject, Versions-class

[Package Biobase version 2.70.0 Index]