validMsg {Biobase}	R Documentation

Conditionally append result to validity message

Description

This function facilitates constructing messages during S4 class validation, and is meant for developer rather than end-user use.

Usage

validMsg(msg, result)

Arguments

msg
A character vector or NULL.

result
Any vector.

Details

This function appends result to msg, but only if result is a character vector.

Author(s)

Martin Morgan <mtmorgan@fhcrc.org>

Examples

msg <- NULL
validMsg(msg, FALSE) # still NULL
msg <- validMsg(msg, "one")
validMsg(msg, "two")

[Package Biobase version 2.70.0 Index]